Fixing warnings on FreeDNS update NOP
Fix based on pull request #31 on github commited by @hank.
_original comments_
When FreeDNS was updating, I was noticing that it would send back an HTTP 200
with the text "Address [IP] has not changed", but ddclient was issuing a
failed status for this and logging a ton of text to syslog. I changed the
function a tiny bit to allow this to be a success status. I also reformatted
the code to all spaces instead of mixed spaces and tabs, but you can take that
or leave it - I don't really care, it was just a total whitespace disaster
before.