Only write the PID file if daemonized
There's not much point in writing a PID file if the user does not want
to fork. (The process executing ddclient already knows the PID, so it
can save it to a file if desired.)

This change is a prerequisite to using Proc::Daemon to fix the buggy
daemonization logic.