Create NEWS.md, seeded with RELEASENOTE's history
Create a document that describes notable changes between released
versions. Populate it with the contents of RELEASENOTE over time, and
delete RELEASENOTE as it is now redundant. (The full announcements can
still be found at https://github.com/ddclient/ddclient/releases.)

The name NEWS.md was chosen because it is the convention, and because
Debian policy says documents of this nature must be installed at
/usr/share/doc/ddclient/NEWS.gz. For details, see:
https://www.debian.org/doc/debian-policy/ch-docs.html#changelog-files-and-release-notes